refactor: change task module structure fix: remove uneccesarry mutable static
