+++ pros/src/task/local.rs
@@ -6,7 +6,7 @@ use spin::Once;
 
 use super::current;
 
-static mut INDEX: AtomicU32 = AtomicU32::new(0);
+static INDEX: AtomicU32 = AtomicU32::new(0);
 
 // Unsafe because you can change the thread local storage while it is being read.
 // This requires you to leak val so that you can be sure it lives the entire task.
@@ -19,6 +19,23 @@ unsafe fn task_local_storage_set<T>(task: pros_sys::task_t, val: &'static T, ind
 unsafe fn task_local_storage_get<T>(task: pros_sys::task_t, index: u32) -> Option<&'static T> {
     let val = pros_sys::pvTaskGetThreadLocalStoragePointer(task, index as _);
     val.cast::<T>().as_ref()
+}
+
+fn fetch_storage() -> &'static RefCell<ThreadLocalStorage> {
+    let current = current();
+
+    // Get the thread local storage for this task.
+    // Creating it if it doesn't exist.
+    // This is safe as long as index 0 of the freeRTOS TLS is never set to any other type.
+    unsafe {
+        task_local_storage_get(current.task, 0).unwrap_or_else(|| {
+            let storage = Box::leak(Box::new(RefCell::new(ThreadLocalStorage {
+                data: BTreeMap::new(),
+            })));
+            task_local_storage_set(current.task, storage, 0);
+            storage
+        })
+    }
 }
 
 struct ThreadLocalStorage {
@@ -38,27 +55,28 @@ impl<T: 'static> LocalKey<T> {
         }
     }
 
+    fn index(&'static self) -> &usize {
+        self.index.call_once(|| INDEX.fetch_add(1, core::sync::atomic::Ordering::Relaxed) as _)
+    }
+
+    pub fn set(&'static self, val: T) {
+        let storage = fetch_storage();
+        let index = *self.index();
+
+        let val = Box::leak(Box::new(val));
+
+        storage.borrow_mut().data.insert(index, NonNull::new((val as *mut T).cast()).unwrap());
+    }
+
     pub fn with<F, R>(&'static self, f: F) -> R
     where
         F: FnOnce(&'static T) -> R,
     {
-        let index = *self.index.call_once(|| unsafe {
-            INDEX.fetch_add(1, core::sync::atomic::Ordering::SeqCst) as usize
-        });
-        let current = current();
+        let storage = fetch_storage();
+        let index = *self.index();
 
-        // Get the thread local storage for this task.
-        // Creating it if it doesn't exist.
-        let storage = unsafe {
-            task_local_storage_get(current.task, 0).unwrap_or_else(|| {
-                let storage = Box::leak(Box::new(RefCell::new(ThreadLocalStorage {
-                    data: BTreeMap::new(),
-                })));
-                task_local_storage_set(current.task, storage, 0);
-                storage
-            })
-        };
-
+        // Make sure that the borrow is dropped if the if does not execute.
+        // This shouldn't be necessary, but caution is good.
         {
             if let Some(val) = storage.borrow_mut().data.get(&index) {
                 return f(unsafe { val.cast().as_ref() });
@@ -66,7 +84,10 @@ impl<T: 'static> LocalKey<T> {
         }
 
         let val = Box::leak(Box::new((self.init)()));
-        storage.borrow_mut().data.insert(index, NonNull::new((val as *mut T).cast::<()>()).unwrap());
+        storage
+            .borrow_mut()
+            .data
+            .insert(index, NonNull::new((val as *mut T).cast::<()>()).unwrap());
         f(val)
     }
 }
